Interface flow
I think this covers the basic flow of the interface - still need to
figure out how to have a function that acts as a while loop to detect
when everything has been added to the database.  Might need to figure
out pyqtSignal or toggle for that.  Going to start incorporating
input_data and server next.